+++ decode_map.go
@@ -18,7 +18,7 @@ var (
 )
 
 var (
-	mapStringInterfacePtrType = reflect.TypeOf((*map[string]interface{})(nil))
+	mapStringInterfacePtrType = reflect.TypeOf((*map[string]any)(nil))
 	mapStringInterfaceType    = mapStringInterfacePtrType.Elem()
 )
 
@@ -48,7 +48,7 @@ func decodeMapValue(d *Decoder, v reflect.Value) error {
 	return d.decodeTypedMapValue(v, n)
 }
 
-func (d *Decoder) decodeMapDefault() (interface{}, error) {
+func (d *Decoder) decodeMapDefault() (any, error) {
 	if d.mapDecoder != nil {
 		return d.mapDecoder(d)
 	}
@@ -134,11 +134,11 @@ func (d *Decoder) decodeMapStringStringPtr(ptr *map[string]string) error {
 }
 
 func decodeMapStringInterfaceValue(d *Decoder, v reflect.Value) error {
-	ptr := v.Addr().Convert(mapStringInterfacePtrType).Interface().(*map[string]interface{})
+	ptr := v.Addr().Convert(mapStringInterfacePtrType).Interface().(*map[string]any)
 	return d.decodeMapStringInterfacePtr(ptr)
 }
 
-func (d *Decoder) decodeMapStringInterfacePtr(ptr *map[string]interface{}) error {
+func (d *Decoder) decodeMapStringInterfacePtr(ptr *map[string]any) error {
 	m, err := d.DecodeMap()
 	if err != nil {
 		return err
@@ -147,7 +147,7 @@ func (d *Decoder) decodeMapStringInterfacePtr(ptr *map[string]interface{}) error
 	return nil
 }
 
-func (d *Decoder) DecodeMap() (map[string]interface{}, error) {
+func (d *Decoder) DecodeMap() (map[string]any, error) {
 	n, err := d.DecodeMapLen()
 	if err != nil {
 		return nil, err
@@ -157,7 +157,7 @@ func (d *Decoder) DecodeMap() (map[string]interface{}, error) {
 		return nil, nil
 	}
 
-	m := make(map[string]interface{}, n)
+	m := make(map[string]any, n)
 
 	for i := 0; i < n; i++ {
 		mk, err := d.DecodeString()
@@ -174,7 +174,7 @@ func (d *Decoder) DecodeMap() (map[string]interface{}, error) {
 	return m, nil
 }
 
-func (d *Decoder) DecodeUntypedMap() (map[interface{}]interface{}, error) {
+func (d *Decoder) DecodeUntypedMap() (map[any]any, error) {
 	n, err := d.DecodeMapLen()
 	if err != nil {
 		return nil, err
@@ -184,7 +184,7 @@ func (d *Decoder) DecodeUntypedMap() (map[interface{}]interface{}, error) {
 		return nil, nil
 	}
 
-	m := make(map[interface{}]interface{}, n)
+	m := make(map[any]any, n)
 
 	for i := 0; i < n; i++ {
 		mk, err := d.decodeInterfaceCond()
@@ -205,7 +205,7 @@ func (d *Decoder) DecodeUntypedMap() (map[interface{}]interface{}, error) {
 
 // DecodeTypedMap decodes a typed map. Typed map is a map that has a fixed type for keys and values.
 // Key and value types may be different.
-func (d *Decoder) DecodeTypedMap() (interface{}, error) {
+func (d *Decoder) DecodeTypedMap() (any, error) {
 	n, err := d.DecodeMapLen()
 	if err != nil {
 		return nil, err