+++ CHANGELOG.md
@@ -63,7 +63,7 @@
 ## v3
 
 - gopkg.in is not supported any more. Update import path to github.com/vmihailenco/msgpack.
-- Msgpack maps are decoded into map[string]interface{} by default.
+- Msgpack maps are decoded into map[string]any by default.
 - EncodeSliceLen is removed in favor of EncodeArrayLen. DecodeSliceLen is removed in favor of
   DecodeArrayLen.
 - Embedded structs are automatically inlined where possible.